1.30 Refactor 4 activities into 4 fragments
